[load_balancing] Move interfaces to src/core/lib/load_balancing
