+++ src/core/ext/filters/client_channel/lb_policy/child_policy_handler.h
@@ -22,11 +22,11 @@
 
 #include "absl/strings/string_view.h"
 
-#include "src/core/ext/filters/client_channel/lb_policy.h"
 #include "src/core/lib/channel/channel_args.h"
 #include "src/core/lib/debug/trace.h"
 #include "src/core/lib/gprpp/orphanable.h"
 #include "src/core/lib/gprpp/ref_counted_ptr.h"
+#include "src/core/lib/load_balancing/lb_policy.h"
 
 namespace grpc_core {
