+++ src/core/ext/filters/client_channel/lb_policy/oob_backend_metric.h
@@ -22,8 +22,8 @@
 #include <memory>
 
 #include "src/core/ext/filters/client_channel/lb_policy/backend_metric_data.h"
-#include "src/core/ext/filters/client_channel/subchannel_interface.h"
 #include "src/core/lib/gprpp/time.h"
+#include "src/core/lib/load_balancing/subchannel_interface.h"
 
 namespace grpc_core {
