+++ src/core/lib/load_balancing/lb_policy.cc
@@ -18,7 +18,7 @@
 
 #include <grpc/support/port_platform.h>
 
-#include "src/core/ext/filters/client_channel/lb_policy.h"
+#include "src/core/lib/load_balancing/lb_policy.h"
 
 #include "src/core/lib/iomgr/closure.h"
 #include "src/core/lib/iomgr/error.h"
