+++ src/core/lib/load_balancing/lb_policy_registry.cc
@@ -16,7 +16,7 @@
 
 #include <grpc/support/port_platform.h>
 
-#include "src/core/ext/filters/client_channel/lb_policy_registry.h"
+#include "src/core/lib/load_balancing/lb_policy_registry.h"
 
 #include <algorithm>
 #include <map>
