+++ test/core/util/BUILD
@@ -293,10 +293,14 @@ grpc_cc_library(
         "//:grpc_client_channel",
         "//:json",
         "//:json_util",
+        "//:lb_policy",
+        "//:lb_policy_factory",
+        "//:lb_policy_registry",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:resolved_address",
         "//:server_address",
+        "//:subchannel_interface",
         "//:time",
         "//:uri_parser",
     ],
